Cross-Validate Mixed-Effects Model

cv() methods for models of class "merMod", fit by the lmer() and glmer() functions in the lme4 package; for models of class "lme" fit by the lme() function in the nlme package; and for models of class "glmmTMB" fit by the glmmTMB() function in the glmmTMB package. The cvMixed() function is meant to be called by cv() methods for mixed-effect models and not directly by the user. It can be used to extend cv() to other classes of mixed-effects models.

Usage

cvMixed(
  model,
  package,
  data = insight::get_data(model),
  criterion = mse,
  k,
  reps = 1,
  seed,
  ncores = 1,
  clusterVariables,
  predict.clusters.args = list(object = model, newdata = data),
  predict.cases.args = list(object = model, newdata = data),
  ...
)

# S3 method for merMod
cv(
  model,
  data = insight::get_data(model),
  criterion = mse,
  k,
  reps = 1,
  seed,
  ncores = 1,
  clusterVariables,
  ...
)

# S3 method for lme
cv(
  model,
  data = insight::get_data(model),
  criterion = mse,
  k,
  reps = 1,
  seed,
  ncores = 1,
  clusterVariables,
  ...
)

# S3 method for glmmTMB
cv(
  model,
  data = insight::get_data(model),
  criterion = mse,
  k,
  reps = 1,
  seed,
  ncores = 1,
  clusterVariables,
  ...
)

Arguments

model

a mixed-effects model object for which a cv() method is available.

package

the name of the package in which mixed-modeling function (or functions) employed resides; used to get the namespace of the package.

data

data frame to which the model was fit (not usually necessary)

criterion

cross-validation ("cost" or lack-of-fit) criterion function of form f(y, yhat) where y is the observed values of the response and yhat the predicted values; the default is mse (the mean-squared error)

k

perform k-fold cross-validation; k may be a number or "loo" or "n" for n-fold (leave-one-out) cross-validation; the default is 10 if cross-validating individual cases and "loo" if cross-validating clusters.

reps

number of times to replicate k-fold CV (default is 1)

seed

for R's random number generator; optional, if not supplied a random seed will be selected and saved; not needed for n-fold cross-validation

ncores

number of cores to use for parallel computations (default is 1, i.e., computations aren't done in parallel)

clusterVariables

a character vector of names of the variables defining clusters for a mixed model with nested random effects; if missing, cross-validation is performed for individual cases rather than for clusters

predict.clusters.args

a list of arguments to be used to predict the whole data set from a mixed model when performing CV on clusters; the first two elements should be model and newdata; see the "Extending the cv package" vignette (vignette("cv-extend", package="cv")).

predict.cases.args

a list of arguments to be used to predict the whole data set from a mixed model when performing CV on cases; the first two elements should be model and newdata; see the "Extending the cv package" vignette (vignette("cv-extend", package="cv")).

...

for cv() methods, to match generic, and for cvMixed(), arguments to be passed to update().

Value

cvMixed(), and functions based on it, such as the methods cv.merMod(), cv.lme(), and cv.glmmTMB(), return objects of class "cv", or, if reps > 1, of class "cvList" (see cv()).

Details

For mixed-effects models, cross-validation can be done by "clusters" or by individual observations. If the former, predictions are based only on fixed effects; if the latter, predictions include the random effects (i.e., are the best linear unbiased predictors or "BLUPS"). Only mixed models with fully nested random effects are supported.

Functions

cvMixed(): not to be called directly

cv(merMod): cv() method

cv(lme): cv() method

cv(glmmTMB): cv() method

See also

Examples

library("lme4")
#> Loading required package: Matrix
# from ?lmer:
(fm1 <- lmer(Reaction ~ Days + (Days | Subject), sleepstudy))
#> Linear mixed model fit by REML ['lmerMod']
#> Formula: Reaction ~ Days + (Days | Subject)
#>    Data: sleepstudy
#> REML criterion at convergence: 1743.628
#> Random effects:
#>  Groups   Name        Std.Dev. Corr
#>  Subject  (Intercept) 24.741       
#>           Days         5.922   0.07
#>  Residual             25.592       
#> Number of obs: 180, groups:  Subject, 18
#> Fixed Effects:
#> (Intercept)         Days  
#>      251.41        10.47  
cv(fm1, clusterVariables="Subject") # LOO CV of clusters
#> n-Fold Cross Validation based on 18 {Subject} clusters
#> cross-validation criterion = 2460.604
#> bias-adjusted cross-validation criterion = 2454.627
#> full-sample criterion = 2251.398 
cv(fm1, seed=447) # 10-fold CV of cases
#> R RNG seed set to 447
#> 10-Fold Cross Validation
#> cross-validation criterion = 869.533
#> bias-adjusted cross-validation criterion = 847.5586
#> full-sample criterion = 549.342 
cv(fm1, clusterVariables="Subject", k=5,
   seed=834, reps=3) # 5-fold CV of clusters, repeated 3 times
#> R RNG seed set to 834
#> R RNG seed set to 448690
#> R RNG seed set to 916534
#> 
#> Replicate 1:
#> 5-Fold Cross Validation based on 18 {Subject} clusters
#> cross-validation criterion = 2458.192
#> bias-adjusted cross-validation criterion = 2434.117
#> full-sample criterion = 2251.398 
#> 
#> Replicate 2:
#> 5-Fold Cross Validation based on 18 {Subject} clusters
#> cross-validation criterion = 2511.651
#> bias-adjusted cross-validation criterion = 2479.989
#> full-sample criterion = 2251.398 
#> 
#> Replicate 3:
#> 5-Fold Cross Validation based on 18 {Subject} clusters
#> cross-validation criterion = 2403.755
#> bias-adjusted cross-validation criterion = 2387.049
#> full-sample criterion = 2251.398 
#> 
#> Average:
#> 5-Fold Cross Validation based on 18 {Subject} clusters
#> cross-validation criterion = 2457.947 (44.04918)
#> bias-adjusted cross-validation criterion = 2433.818 (37.94423)
#> full-sample criterion = 2251.398 

library(nlme)
#> 
#> Attaching package: ‘nlme’
#> The following object is masked from ‘package:lme4’:
#> 
#>     lmList
#> The following object is masked from ‘package:cv’:
#> 
#>     getResponse
# from ?lme
(fm2 <- lme(distance ~ age + Sex, data = Orthodont,
            random = ~ 1))
#> Linear mixed-effects model fit by REML
#>   Data: Orthodont 
#>   Log-restricted-likelihood: -218.7563
#>   Fixed: distance ~ age + Sex 
#> (Intercept)         age   SexFemale 
#>  17.7067130   0.6601852  -2.3210227 
#> 
#> Random effects:
#>  Formula: ~1 | Subject
#>         (Intercept) Residual
#> StdDev:    1.807425 1.431592
#> 
#> Number of Observations: 108
#> Number of Groups: 27 
cv(fm2) # LOO CV of cases
#> R RNG seed set to 765199
#> 10-Fold Cross Validation
#> cross-validation criterion = 2.666103
#> bias-adjusted cross-validation criterion = 2.589541
#> full-sample criterion = 1.582435 
cv(fm2, clusterVariables="Subject", k=5, seed=321) # 5-fold CV of clusters
#> R RNG seed set to 321
#> 5-Fold Cross Validation based on 27 {Subject} clusters
#> cross-validation criterion = 5.875411
#> bias-adjusted cross-validation criterion = 5.780695
#> full-sample criterion = 5.017326 

library("glmmTMB")
# from ?glmmTMB
(m1 <- glmmTMB(count ~ mined + (1|site),
               zi=~mined,
               family=poisson, data=Salamanders))
#> Formula:          count ~ mined + (1 | site)
#> Zero inflation:         ~mined
#> Data: Salamanders
#>       AIC       BIC    logLik  df.resid 
#> 1908.4695 1930.8080 -949.2348       639 
#> Random-effects (co)variances:
#> 
#> Conditional model:
#>  Groups Name        Std.Dev.
#>  site   (Intercept) 0.28    
#> 
#> Number of obs: 644 / Conditional model: site, 23
#> 
#> Fixed Effects:
#> 
#> Conditional model:
#> (Intercept)      minedno  
#>      0.0879       1.1419  
#> 
#> Zero-inflation model:
#> (Intercept)      minedno  
#>       1.139       -1.736  
cv(m1, seed=97816, k=5) # 5-fold CV of clusters
#> R RNG seed set to 97816
#> 5-Fold Cross Validation
#> cross-validation criterion = 6.110784
#> bias-adjusted cross-validation criterion = 6.075783
#> full-sample criterion = 5.79783 
cv(m1, seed=34506, k=5) # 5-fold CV of cases
#> R RNG seed set to 34506
#> 5-Fold Cross Validation
#> cross-validation criterion = 6.058988
#> bias-adjusted cross-validation criterion = 6.032337
#> full-sample criterion = 5.79783