Cross-Validate a Model-Selection Procedure

A generic function to cross-validate a model-selection procedure, along with a procedure that applies the stepAIC() model-selection function in the MASS package, and a procedure for selecting predictor and response transformations in regression, which uses the powerTransform() function in the car package.

Usage

cvSelect(
  procedure,
  data,
  k = 10,
  reps = 1,
  save.coef = k <= 10,
  seed,
  ncores = 1,
  ...
)

selectStepAIC(
  data,
  indices,
  model,
  criterion = mse,
  AIC = TRUE,
  save.coef = TRUE,
  ...
)

selectTrans(
  data,
  indices,
  save.coef = TRUE,
  model,
  criterion = mse,
  predictors,
  response,
  family = c("bcPower", "bcnPower", "yjPower", "basicPower"),
  family.y = c("bcPower", "bcnPower", "yjPower", "basicPower"),
  rounded = TRUE,
  ...
)

compareFolds(object, digits = 3, ...)

Arguments

procedure

a model-selection procedure function (see Details).

data

full data frame for model selection.

k

perform k-fold cross-validation (default is 10); k may be a number or "loo" or "n" for n-fold (leave-one-out) cross-validation.

reps

number of times to replicate k-fold CV (default is 1)

save.coef

save the coefficients from the selected models? (default is TRUE if k is 10 or smaller, FALSE otherwise)

seed

for R's random number generator; not used for n-fold cross-validation.

ncores

number of cores to use for parallel computations (default is 1, i.e., computations aren't done in parallel)

...

arguments to be passed to procedure().

indices

indices of cases in data defining the current fold.

model

a regression model object fit to data.

criterion

a CV criterion function.

AIC

if TRUE (the default) use the AIC as the model-selection criterion; if FALSE, use the BIC. The k argument to stepAIC() is set accordingly (note that this is distinct from the number of folds k).

predictors

character vector of names of the predictors in the model to transform; if missing, no predictors will be transformed.

response

name of the response variable; if missing, the response won't be transformed.

family

transformation family for the predictors, one of "bcPower", "bcnPower", "yjPower", "basicPower", with "bcPower" as the default. These are the names of transformation functions in the car package; see bcPower

family.y

transformation family for the response, with "bcPower" as the default.

rounded

if TRUE (the default) use nicely rounded versions of the estimated transformation parameters (see bcPower).

object

an object of class "cvSelect".

digits

significant digits for printing coefficients, (default 3).

Value

An object of class "cvSelect", inheriting from class "cv", with the averaged CV criterion ("CV crit"), the adjusted average CV criterion ("adj CV crit"), the criterion for the model applied to the full data ("full crit"), the number of folds ("k"), the seed for R's random-number generator ("seed"), and (optionally) a list of coefficients (or, in the case of selectTrans(), estimated transformation parameters) for the selected models for each fold ("coefficients"). If reps > 1, then an object of class c("cvSelectList", "cvList") is returned, which is literally a list of c("cvSelect", "cv") objects.

Details

The model-selection function supplied as the procedure argument to cvSelect() should accept the following arguments:

data

set to the data argument to cvSelect().

indices

the indices of the rows of data defining the current fold; if missing, the model-selection procedure is applied to the full data.

other arguments

to be passed via ... from cvSelect().

procedure() should return a two-element vector with the result of applying a cross-validation criterion to the cases in the current fold for the model deleting that fold, and to all of the cases again for the model deleting the current fold.

When the indices argument is missing, procedure() returns the cross-validation criterion for all of the cases based on the model fit to all of the cases.

For examples of model-selection functions for the procedure argument, see the code for selectStepAIC() and for selectTrans().

For additional information, see the "Cross-validation of regression models" vignette (vignette("cv", package="cv")) and the "Extending the cv package" vignette (vignette("cv-extend", package="cv")).

Functions

cvSelect(): apply cross-validation to a model-selection procedure.

selectStepAIC(): select a model using the stepAIC() function in the MASS package.

selectTrans(): select transformations of the predictors and response.

compareFolds(): print the coefficients from the selected models for the several folds.

See also

Examples

data("Auto", package="ISLR2")
m.auto <- lm(mpg ~ . - name - origin, data=Auto)
cvSelect(selectStepAIC, Auto, seed=123, model=m.auto)
#> R RNG seed set to 123
#> 10-Fold Cross Validation
#> cross-validation criterion = 11.87795
#> bias-adjusted cross-validation criterion = 11.8753
#> full-sample criterion = 11.65549 
cvSelect(selectStepAIC, Auto, seed=123, model=m.auto,
         AIC=FALSE, k=5, reps=3) # via BIC
#> R RNG seed set to 123
#> R RNG seed set to 319729
#> R RNG seed set to 889422
#> 
#> Replicate 1:
#> 5-Fold Cross Validation
#> cross-validation criterion = 11.91891
#> bias-adjusted cross-validation criterion = 11.8895
#> full-sample criterion = 11.65549 
#> 
#> Replicate 2:
#> 5-Fold Cross Validation
#> cross-validation criterion = 11.80601
#> bias-adjusted cross-validation criterion = 11.78946
#> full-sample criterion = 11.65549 
#> 
#> Replicate 3:
#> 5-Fold Cross Validation
#> cross-validation criterion = 11.81343
#> bias-adjusted cross-validation criterion = 11.79576
#> full-sample criterion = 11.65549 
#> 
#> Average:
#> 5-Fold Cross Validation
#> cross-validation criterion = 11.86432 (0.06311657)
#> bias-adjusted cross-validation criterion = 11.84105 (0.05600268)
#> full-sample criterion = 11.65549 

data("Prestige", package="carData")
m.pres <- lm(prestige ~ income + education + women,
             data=Prestige)
cvt <- cvSelect(selectTrans, data=Prestige, model=m.pres, seed=123,
                predictors=c("income", "education", "women"),
                response="prestige", family="yjPower")
#> R RNG seed set to 123
cvt
#> 10-Fold Cross Validation
#> cross-validation criterion = 58.68193
#> bias-adjusted cross-validation criterion = 58.26258
#> full-sample criterion = 50.60016 
compareFolds(cvt)
#>         lam.education lam.income lam.women lamda.y
#> Fold 1          1.000      0.330     0.330       1
#> Fold 2          1.000      0.330     0.330       1
#> Fold 3          1.000      0.330     0.000       1
#> Fold 4          1.000      0.330     0.330       1
#> Fold 5          1.000      0.330     0.000       1
#> Fold 6          1.000      0.330     0.000       1
#> Fold 7          1.000      0.000     0.000       1
#> Fold 8          1.000      0.330     0.000       1
#> Fold 9          1.000      0.500     0.330       1
#> Fold 10         1.000      0.330     0.159       1
cv(m.pres, seed=123)
#> R RNG seed set to 123
#> 10-Fold Cross Validation
#> method: Woodbury
#> criterion: mse
#> cross-validation criterion = 66.91929
#> bias-adjusted cross-validation criterion = 66.45514
#> full-sample criterion = 59.15265