Extract Response Variable(s)

Generic function to extract the response variable(s) from a fitted model.

Usage

getResponse(model, ...)

# S3 method for default
getResponse(model, ...)

Arguments

model

a fitted model

...

additional parameters for specific methods

Value

a vector or matrix containing the values of response variable(s)

Details

The supplied default method returns the model$y component of the model object if it exists and otherwise the result of model.response(model.frame(model)), checking in either case whether the result is numeric.

Methods (by class)

getResponse(default): default method

Examples

    fit <- lm(cbind(hp, mpg) ~ gear, mtcars)
    getResponse(fit)
#>                      hp  mpg
#> Mazda RX4           110 21.0
#> Mazda RX4 Wag       110 21.0
#> Datsun 710           93 22.8
#> Hornet 4 Drive      110 21.4
#> Hornet Sportabout   175 18.7
#> Valiant             105 18.1
#> Duster 360          245 14.3
#> Merc 240D            62 24.4
#> Merc 230             95 22.8
#> Merc 280            123 19.2
#> Merc 280C           123 17.8
#> Merc 450SE          180 16.4
#> Merc 450SL          180 17.3
#> Merc 450SLC         180 15.2
#> Cadillac Fleetwood  205 10.4
#> Lincoln Continental 215 10.4
#> Chrysler Imperial   230 14.7
#> Fiat 128             66 32.4
#> Honda Civic          52 30.4
#> Toyota Corolla       65 33.9
#> Toyota Corona        97 21.5
#> Dodge Challenger    150 15.5
#> AMC Javelin         150 15.2
#> Camaro Z28          245 13.3
#> Pontiac Firebird    175 19.2
#> Fiat X1-9            66 27.3
#> Porsche 914-2        91 26.0
#> Lotus Europa        113 30.4
#> Ford Pantera L      264 15.8
#> Ferrari Dino        175 19.7
#> Maserati Bora       335 15.0
#> Volvo 142E          109 21.4