Cross-Validate Regression Models

A parallelized generic k-fold (including n-fold, i.e., leave-one-out) cross-validation function, with a default method, and specific methods for linear and generalized-linear models that can be much more computationally efficient.

Usage

# S3 method for merMod
cv(
  model,
  data = insight::get_data(model),
  criterion = mse,
  k = nrow(clusters),
  seed,
  parallel = FALSE,
  ncores = parallelly::availableCores(logical = FALSE),
  clusterVariables,
  ...
)

cv(model, data, criterion, k, seed, ...)

# S3 method for default
cv(
  model,
  data = insight::get_data(model),
  criterion = mse,
  k = 10,
  seed,
  parallel = FALSE,
  ncores = parallelly::availableCores(logical = FALSE),
  method = NULL,
  ...
)

# S3 method for cv
print(x, digits = getOption("digits"), ...)

# S3 method for lm
cv(
  model,
  data = insight::get_data(model),
  criterion = mse,
  k = 10,
  seed,
  method = c("auto", "hatvalues", "Woodbury", "naive"),
  parallel = FALSE,
  ncores = parallelly::availableCores(logical = FALSE),
  ...
)

# S3 method for glm
cv(
  model,
  data = insight::get_data(model),
  criterion = mse,
  k = 10,
  seed,
  method = c("exact", "hatvalues", "Woodbury"),
  parallel = FALSE,
  ncores = parallelly::availableCores(logical = FALSE),
  ...
)

Arguments

model

a regression model object (see Details).

data

data frame to which the model was fit (not usually necessary)

criterion

cross-validation criterion function of form f(y, yhat) where y is the observed values of the response and yhat the predicted values; the default is mse (the mean-squared error)

k

perform k-fold cross-validation (default is 10); k may be a number or "loo" or "n" for n-fold (leave-one-out) cross-validation.

seed

for R's random number generator; optional, if not supplied a random seed will be selected and saved; not needed for n-fold cross-validation

parallel

do computations in parallel? (default is FALSE)

ncores

number of cores to use for parallel computations (default is number of physical cores detected)

clusterVariables

a character vector of names of the variables defining clusters for a mixed model with nested random effects

...

to match generic

method

computational method to apply to a linear (i.e. "lm") model or to a generalized linear (i.e., "glm") model. See Details for an explanation of the available options.

x

a cv object to be printed

digits

significant digits for printing, default taken from the "digits" option

Value

a "cv" object with the cross-validation criterion averaged across the folds, the bias-adjusted averaged CV criterion, the criterion applied to the model fit to the full data set, and the initial value of R's RNG seed. Some methods return a subset of these components and may add additional information.

Details

The default method uses update() to refit the model to each fold, and should work if there are appropriate update() and predict() methods and if the default method for getResponse() works or if a getResponse() method is supplied.

The "lm" and "glm" methods can use much faster computational algorithms, as selected by the method argument. The linear-model method accommodates weighted linear models.

For both classes of models, for the leave-one-out (n-fold) case, fitted values for the folds can be computed from the hat-values via method="hatvalues" without refitting the model; for GLMs, this method is approximate, for LMs it is exact.

Again for both classes of models, when more than one case is omitted in each fold, fitted values may be obtained without refitting the model by exploiting the Woodbury matrix identity via method="Woodbury". As for hatvalues, this method is exact for LMs and approximate for GLMs.

The default for linear models is method="auto", which is equivalent to method="hatvalues" for n-fold cross-validation and method="Woodbury" otherwise; method="naive" refits the model via update() and is generally much slower. The default for generalized linear models is method="exact", which employs update().

Methods (by class)

cv(merMod): merMod method

cv(default): default method

cv(lm): "lm" method

cv(glm): "glm" method

Methods (by generic)

print(cv): print() method

Examples

data("Auto", package="ISLR2")
m.auto <- lm(mpg ~ horsepower, data=Auto)
cv(m.auto,  k="loo")
#> n-Fold Cross Validation
#> method: hatvalues
#> criterion: mse
#> cross-validation criterion = 24.23151
cv(m.auto, seed=1234)
#> R RNG seed set to 1234
#> 10-Fold Cross Validation
#> method: Woodbury
#> criterion: mse
#> cross-validation criterion = 24.3794
#> bias-adjusted cross-validation criterion = 24.35646
#> full-sample criterion = 23.94366 

data("Caravan", package="ISLR2")
m.caravan <- glm(Purchase ~ ., data=Caravan[1:2500, ], family=binomial)
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
cv(m.caravan, k=5, criterion=BayesRule, seed=123)
#> R RNG seed set to 123
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: prediction from rank-deficient fit; attr(*, "non-estim") has doubtful cases
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: prediction from rank-deficient fit; attr(*, "non-estim") has doubtful cases
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: prediction from rank-deficient fit; attr(*, "non-estim") has doubtful cases
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: prediction from rank-deficient fit; attr(*, "non-estim") has doubtful cases
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: prediction from rank-deficient fit; attr(*, "non-estim") has doubtful cases
#> 5-Fold Cross Validation
#> method: exact
#> criterion: BayesRule
#> cross-validation criterion = 0.0648
#> bias-adjusted cross-validation criterion = 0.06512
#> full-sample criterion = 0.0576